// libs/dots-system/src/types/use-history.ts
import { FC } from 'react';

export interface HistoryItem {
  path: string;
  title: string;
  Component: FC;
  componentProps?: {
    [key: string]: unknown;
  };
  width?: 'sm' | 'md' | 'lg' | 'xl';
}

export interface HistoryState {
  paths: string[];
  past: HistoryItem[];
  present: HistoryItem | null;
  future: HistoryItem[];
}

export interface HistoryAction {
  type: string;
  payload?: HistoryItem | number | string | null;
}